moved code into separate modules and dirs
