+++ src/lib.rs
@@ -1,12 +1,5 @@
 pub mod engine;
-pub mod zone;
-pub mod tab;
-pub mod net;
-pub mod tick;
-pub mod config;
-pub mod instance;
-pub mod event;
-pub mod errors;
 pub mod viewport;
+pub mod net;
 
 pub use engine::*;
